Entrez E-utilities
Intial commit: handling the monostrosities of data types they defined.
encoding/xml is not used to fill structs because the defined data types
do not easily parse into structs - relative ordering of a number of
mutually exlusive types is not doable with encoding/xml. In addition to
this, encoding/xml fails silently on unknown tags and given the way NCBI
randomly change formats (as noted in the internal documentation here) it
would be unwise to trust data from returned streams. Also, there is no
need to re-encode and I do not want to give the impression that these
types should be used that way.

In future, please use JSON.

Note that the currently defined data types are not yet stable.