BLAST support
This was a joy to write. Primarily because of the html and xml packages.
They just work.

TODO tip html package is forked into the repository until Go1.1 is
released. It will not pass its tests due to dependencies on standard
library features that are not in Go1.0.3. Remove it when Go1.1 is
released.